src/val-service/requirements.txt
google-cloud-pubsub==2.23.0
google-cloud-secret-manager==2.20.0
langchain-google-genai==2.1.0
langgraph==0.3.30
langgraph-checkpoint-postgres==2.0.15
psycopg[binary,pool]==3.2.0
psycopg2-binary==2.9.9
weaviate-client==4.9.0
